+++ rust/operator-binary/src/controller.rs
@@ -26,11 +26,9 @@ use stackable_operator::{
     },
     client::Client,
     cluster_resources::{ClusterResourceApplyStrategy, ClusterResources},
-    commons::{
-        authentication::oidc::AuthenticationProvider,
-        product_image_selection::ResolvedProductImage, rbac::build_rbac_resources,
-    },
+    commons::{product_image_selection::ResolvedProductImage, rbac::build_rbac_resources},
     config::fragment,
+    crd::{authentication::oidc, git_sync},
     k8s_openapi::{
         DeepMerge,
         api::{
@@ -52,6 +50,7 @@ use stackable_operator::{
     kvp::{Label, Labels, ObjectLabels},
     logging::controller::ReconcilerError,
     memory::{BinaryMultiple, MemoryQuantity},
+    product_config_utils::env_vars_from_rolegroup_config,
     product_logging::{
         self,
         framework::{
@@ -77,8 +76,9 @@ use crate::{
     OPERATOR_NAME,
     config::{
         self, JVM_SECURITY_PROPERTIES_FILE, NIFI_BOOTSTRAP_CONF, NIFI_CONFIG_DIRECTORY,
-        NIFI_PROPERTIES, NIFI_STATE_MANAGEMENT_XML, NifiRepository, build_bootstrap_conf,
-        build_nifi_properties, build_state_management_xml, validated_product_config,
+        NIFI_PROPERTIES, NIFI_PYTHON_WORKING_DIRECTORY, NIFI_STATE_MANAGEMENT_XML, NifiRepository,
+        build_bootstrap_conf, build_nifi_properties, build_state_management_xml,
+        validated_product_config,
     },
     crd::{
         APP_NAME, BALANCE_PORT, BALANCE_PORT_NAME, Container, CurrentlySupportedListenerClasses,
@@ -109,6 +109,7 @@ pub const NIFI_CONTROLLER_NAME: &str = "nificluster";
 pub const NIFI_FULL_CONTROLLER_NAME: &str = concatcp!(NIFI_CONTROLLER_NAME, '.', OPERATOR_NAME);
 
 const DOCKER_IMAGE_BASE_NAME: &str = "nifi";
+const LOG_VOLUME_NAME: &str = "log";
 
 pub struct Ctx {
     pub client: Client,
@@ -257,6 +258,9 @@ pub enum Error {
 
     #[snafu(display("failed to resolve and merge config for role and role group"))]
     FailedToResolveConfig { source: crate::crd::Error },
+
+    #[snafu(display("invalid git-sync specification"))]
+    InvalidGitSyncSpec { source: git_sync::v1alpha1::Error },
 
     #[snafu(display("vector agent is enabled but vector aggregator ConfigMap is missing"))]
     VectorAggregatorConfigMapMissing,
@@ -509,6 +513,16 @@ pub async fn reconcile_nifi(
                 .merged_config(&NifiRole::Node, rolegroup_name)
                 .context(FailedToResolveConfigSnafu)?;
 
+            let git_sync_resources = git_sync::v1alpha1::GitSyncResources::new(
+                &nifi.spec.cluster_config.custom_components_git_sync,
+                &resolved_product_image,
+                &env_vars_from_rolegroup_config(rolegroup_config),
+                &[],
+                LOG_VOLUME_NAME,
+                &merged_config.logging.for_container(&Container::GitSync),
+            )
+            .context(InvalidGitSyncSpecSnafu)?;
+
             let rg_service =
                 build_node_rolegroup_service(nifi, &resolved_product_image, &rolegroup)?;
 
@@ -531,6 +545,7 @@ pub async fn reconcile_nifi(
                 rolegroup_config,
                 &merged_config,
                 &proxy_hosts,
+                &git_sync_resources,
             )
             .await?;
 
@@ -555,6 +570,7 @@ pub async fn reconcile_nifi(
                 rolling_upgrade_supported,
                 replicas,
                 &rbac_sa.name_any(),
+                &git_sync_resources,
             )
             .await?;
 
@@ -716,6 +732,7 @@ async fn build_node_rolegroup_config_map(
     rolegroup_config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
     merged_config: &NifiConfig,
     proxy_hosts: &str,
+    git_sync_resources: &git_sync::v1alpha1::GitSyncResources,
 ) -> Result<ConfigMap> {
     tracing::debug!("building rolegroup configmaps");
 
@@ -784,6 +801,7 @@ async fn build_node_rolegroup_config_map(
                     })?
                     .clone(),
                 resolved_product_image.product_version.as_ref(),
+                git_sync_resources,
             )
             .with_context(|_| BuildProductConfigSnafu {
                 rolegroup: rolegroup.clone(),
@@ -893,6 +911,7 @@ async fn build_node_rolegroup_statefulset(
     rolling_update_supported: bool,
     replicas: Option<i32>,
     service_account_name: &str,
+    git_sync_resources: &git_sync::v1alpha1::GitSyncResources,
 ) -> Result<StatefulSet> {
     tracing::debug!("Building statefulset");
     let role_group = role.role_groups.get(&rolegroup_ref.role_group);
@@ -960,9 +979,11 @@ async fn build_node_rolegroup_statefulset(
     }
 
     if let NifiAuthenticationConfig::Oidc { oidc, .. } = authentication_config {
-        env_vars.extend(AuthenticationProvider::client_credentials_env_var_mounts(
-            oidc.client_credentials_secret_ref.clone(),
-        ));
+        env_vars.extend(
+            oidc::v1alpha1::AuthenticationProvider::client_credentials_env_var_mounts(
+                oidc.client_credentials_secret_ref.clone(),
+            ),
+        );
     }
 
     env_vars.extend(authorization_config.get_env_vars());
@@ -1083,7 +1104,7 @@ async fn build_node_rolegroup_statefulset(
         .context(AddVolumeMountSnafu)?
         .add_volume_mount("sensitiveproperty", "/stackable/sensitiveproperty")
         .context(AddVolumeMountSnafu)?
-        .add_volume_mount("log", STACKABLE_LOG_DIR)
+        .add_volume_mount(LOG_VOLUME_NAME, STACKABLE_LOG_DIR)
         .context(AddVolumeMountSnafu)?
         .add_volume_mount(TRUSTSTORE_VOLUME_NAME, STACKABLE_SERVER_TLS_DIR)
         .context(AddVolumeMountSnafu)?
@@ -1159,7 +1180,7 @@ async fn build_node_rolegroup_statefulset(
         .context(AddVolumeMountSnafu)?
         .add_volume_mount("log-config", STACKABLE_LOG_CONFIG_DIR)
         .context(AddVolumeMountSnafu)?
-        .add_volume_mount("log", STACKABLE_LOG_DIR)
+        .add_volume_mount(LOG_VOLUME_NAME, STACKABLE_LOG_DIR)
         .context(AddVolumeMountSnafu)?
         .add_volume_mount(TRUSTSTORE_VOLUME_NAME, STACKABLE_SERVER_TLS_DIR)
         .context(AddVolumeMountSnafu)?
@@ -1212,8 +1233,30 @@ async fn build_node_rolegroup_statefulset(
             .context(AddVolumeMountSnafu)?;
     }
 
+    let volume_name = "nifi-python-working-directory".to_string();
+    pod_builder
+        .add_empty_dir_volume(&volume_name, None)
+        .context(AddVolumeSnafu)?;
+    container_nifi
+        .add_volume_mount(&volume_name, NIFI_PYTHON_WORKING_DIRECTORY)
+        .context(AddVolumeMountSnafu)?;
+
+    container_nifi
+        .add_volume_mounts(git_sync_resources.git_content_volume_mounts.to_owned())
+        .context(AddVolumeMountSnafu)?;
+
     // We want to add nifi container first for easier defaulting into this container
     pod_builder.add_container(container_nifi.build());
+
+    for container in git_sync_resources.git_sync_containers.iter().cloned() {
+        pod_builder.add_container(container);
+    }
+    for container in git_sync_resources.git_sync_init_containers.iter().cloned() {
+        pod_builder.add_init_container(container);
+    }
+    pod_builder
+        .add_volumes(git_sync_resources.git_content_volumes.to_owned())
+        .context(AddVolumeSnafu)?;
 
     if let Some(ContainerLogConfig {
         choice:
@@ -1252,7 +1295,7 @@ async fn build_node_rolegroup_statefulset(
                     product_logging::framework::vector_container(
                         resolved_product_image,
                         "config",
-                        "log",
+                        LOG_VOLUME_NAME,
                         merged_config.logging.containers.get(&Container::Vector),
                         ResourceRequirementsBuilder::new()
                             .with_cpu_request("250m")
@@ -1318,7 +1361,7 @@ async fn build_node_rolegroup_statefulset(
         })
         .context(AddVolumeSnafu)?
         .add_empty_dir_volume(
-            "log",
+            LOG_VOLUME_NAME,
             // Set volume size to higher than theoretically necessary to avoid running out of disk space as log rotation triggers are only checked by Logback every 5s.
             Some(
                 MemoryQuantity {
@@ -1413,7 +1456,7 @@ async fn build_node_rolegroup_statefulset(
                 ),
                 ..LabelSelector::default()
             },
-            service_name: rolegroup_ref.object_name(),
+            service_name: Some(rolegroup_ref.object_name()),
             template: pod_template,
             update_strategy: Some(StatefulSetUpdateStrategy {
                 type_: if rolling_update_supported {